+++ src/bragerone/__init__.py
@@ -1,5 +1,5 @@
 # src/bragerone/__init__.py
-from importlib.metadata import version, PackageNotFoundError
+from importlib.metadata import PackageNotFoundError, version
 
 try:
     __version__ = version("py-bragerone")